refactor: change .env and package #34
